Collapse BluetoothManager into ReliaBLEManager (Step 4 of 5, #10)
dropping the indirection layer now that the actor work is done directly.

Move BluetoothState/AuthorizationError types and the actor-isolated
state snapshot onto ReliaBLEManager, delete BluetoothManager.swift,
and update docs/tests to match the new public surface.
